requirements.txt
google-genai
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
PyYAML==6.0.2
tqdm==4.66.5
pandas==2.2.3
